+++ test/thirdpartyemailpassword/override.test.js
@@ -171,12 +171,21 @@ describe(`overrideTest: ${printPath("[test/thirdpartyemailpassword/override.test
                         apis: (oI) => {
                             return {
                                 ...oI,
-                                signInUpPOST: async (input) => {
-                                    let response = await oI.signInUpPOST(input);
+                                emailPasswordSignInPOST: async (input) => {
+                                    let response = await oI.emailPasswordSignInPOST(input);
                                     if (response.status === "OK") {
                                         user = response.user;
-                                        newUser = response.createdNewUser;
-                                        type = response.type;
+                                        newUser = false;
+                                        type = "emailpassword";
+                                    }
+                                    return response;
+                                },
+                                emailPasswordSignUpPOST: async (input) => {
+                                    let response = await oI.emailPasswordSignUpPOST(input);
+                                    if (response.status === "OK") {
+                                        user = response.user;
+                                        newUser = true;
+                                        type = "emailpassword";
                                     }
                                     return response;
                                 },
@@ -419,11 +428,25 @@ describe(`overrideTest: ${printPath("[test/thirdpartyemailpassword/override.test
                         apis: (oI) => {
                             return {
                                 ...oI,
-                                signInUpPOST: async (input) => {
-                                    let response = await oI.signInUpPOST(input);
+                                emailPasswordSignInPOST: async (input) => {
+                                    let response = await oI.emailPasswordSignInPOST(input);
                                     user = response.user;
-                                    newUser = response.createdNewUser;
-                                    type = response.type;
+                                    newUser = false;
+                                    type = "emailpassword";
+                                    if (newUser) {
+                                        throw {
+                                            error: "signup error",
+                                        };
+                                    }
+                                    throw {
+                                        error: "signin error",
+                                    };
+                                },
+                                emailPasswordSignUpPOST: async (input) => {
+                                    let response = await oI.emailPasswordSignUpPOST(input);
+                                    user = response.user;
+                                    newUser = true;
+                                    type = "emailpassword";
                                     if (newUser) {
                                         throw {
                                             error: "signup error",